Well, solved part A, but part B is elusive
Clearly there is some more closed form way to calculate what's going
on with the "circle of marbles" thing, but I'm not sure what it is.
As in, really really really not sure what it is. There's some kind of
nearly geometric series going on (nice at the beginning, but then falls
apart once the lose-at-a-7 thing happens). But I cannot get my head
wrapped around what that is supposed to be. :/