Finished day 11
There was a fun "easy" optimization on the part B of this particular
problem, which was pretty interesting. I tried (for fun) doing the
simple (but sloooow) way by just iterating over sizes outside the x/y
loop, and after letting it run up to size 50 or so (which took many
minutes), I realized a simple way to fix that issue, which is the
solution that I now have.