Day 15 complete
I had a variety of issues getting this to work properly in the first
place, but once that was worked out, I basically added a binary search
on damage to the second part and finished part B. Several examples are
in the input this time, simply because I wanted to run several to figure
out where things were going wrong. Yay, subtle ordering bugs!