Did day 16
The first part wasn't too bad, although I had a fun python reference bug
which took me a minute or two to figure out. The second part was also
not too bad, although I carefully got a couple of similar maps confused
when I was putting that together, and got stuck in an infinite loop.
That was rather embarrasing. But I really liked how simple the final
opcode evaluator was, and once I got there, things worked on the first
try. Yay for something going right!