Finished day 17
Had a couple of interesting initial bugs on this one, but got them
sorted out pretty quickly by adding some better debug printing so that I
could see what was going on. Tricks that I think paid off: using a set
of open points instead of a list (to get deduping), and always taking
the largest Y value from the open set, so that we could solve things
filling "bottom up" as it were to get consistent results for reasonably
minimal effort.