Finished Day 18
This is curiously much like an earlier problem (the pots), but it
has a steady state of multiple entries. Kinda fun to figure out how
that all works in a quasi-reasonable way, although my threshold for
deciding things were settled was actually pretty close to not checking
enough! (My threshold was 100, but the pattern is actually 84 in my
case). At any rate, found it, and it was all good.