Day 20 initial
Don't have a solution yet, but have the building block that I think I
want to generate all of the "walk" strings that the RE indicates.
Hopefully that won't be computationally expensive to come up with. Next
up, I need to convert that into the actual reachability graph, and then
calculate the BFS walk to get to all locations and track what the
farthest place to get is.