Finished Day 20
Got my graph traversal to to work, although it was critical to do the
optimization to skip sections of the recursive descent that have an
identical node and position to other walks of the tree -- apparently
lots of that is encoded into the RE in some kind of clever way. Great.
At any rate, solving the second half was a trivial extra bit of data to
get out of my solution to the first part, so we're all done.