Finished day 21
Calculating up the sim was the way to go for part 1, since that just
empties out the answer we need by examining register 3. To do the second
part, I augmented the instruction set to include integer division, which
speeds the entire cycle processing by...well, a lot. Orders of magnitude
less instructions are required. One line change, and the cycle system
was easy enough to crack.