Day 23 progress, maybe?
Tried out an idea that was correct, but way too slow to be of any
utility. That gave me an idea of something else to try, but I'm running
into issues with how manhattan distance and euclidian distance work that
are making my math not turn out how I'd expect. More work will be
required to get my head to manage this stuff. :/