Day 23 b, more progress
...or, at least, more code. I've gotten to the point that my
intersection point generator is working (I think), but unfortunately it
appears to be the case that even the intersection between the smallest
radius nb that I have of interest and its neighbor on just a plane is
ending up with millions of entries at least (and was still counting up
when I stopped it). Needless to say, this approach isn't going to work.