Day 24a done
I had a minor error in my fight system that would allow defeated armies
to still fight. Oops. Fixed that, and ended as soon as I ran into a
series of cases in a row where one army didn't change, and that was the
ticket. Now in to part B!